// src/tcgdb/modules/tcgdb/pokemon/price/change/monthly/dto/tcgdb.pokemon.price.change.monthly.dto.ts
export class TCGdbPokemonPricesChangeMonthlyDTO {
    tcgdbPokemonPricesChangeMonthly: TCGdbPokemonPriceChangeMonthlyDTO[];
}

export class TCGdbPokemonPriceChangeMonthlyDTO {
    tcgdbPokemonPriceChangeMonthlyId: string;
    tcgdbPokemonCardId: string;
    tcgdbPokemonPriceChangeMonthlyTCGPlayerId: number;
    tcgdbPokemonPriceChangeMonthlySetCode: string;
    tcgdbPokemonPriceChangeMonthlyCurrentLowPrice: number;
    tcgdbPokemonPriceChangeMonthlyPreviousLowPrice: number;
    tcgdbPokemonPriceChangeMonthlyLowPriceDifference: number;
    tcgdbPokemonPriceChangeMonthlyLowPricePercentage: number;
    tcgdbPokemonPriceChangeMonthlyCurrentMidPrice: number;
    tcgdbPokemonPriceChangeMonthlyPreviousMidPrice: number;
    tcgdbPokemonPriceChangeMonthlyMidPriceDifference: number;
    tcgdbPokemonPriceChangeMonthlyMidPricePercentage: number;
    tcgdbPokemonPriceChangeMonthlyCurrentHighPrice: number;
    tcgdbPokemonPriceChangeMonthlyPreviousHighPrice: number;
    tcgdbPokemonPriceChangeMonthlyHighPriceDifference: number;
    tcgdbPokemonPriceChangeMonthlyHighPricePercentage: number;
    tcgdbPokemonPriceChangeMonthlyCurrentMarketPrice: number;
    tcgdbPokemonPriceChangeMonthlyPreviousMarketPrice: number;
    tcgdbPokemonPriceChangeMonthlyMarketPriceDifference: number;
    tcgdbPokemonPriceChangeMonthlyMarketPricePercentage: number;
    tcgdbPokemonPriceChangeMonthlySubTypeName: string;
}

export class CreateTCGdbPokemonPriceChangeMonthlyDTO {
    tcgdbPokemonCardId: string;
    tcgdbPokemonPriceChangeMonthlyTCGPlayerId: number;
    tcgdbPokemonPriceChangeMonthlySetCode: string;
    tcgdbPokemonPriceChangeMonthlyCurrentLowPrice: number;
    tcgdbPokemonPriceChangeMonthlyPreviousLowPrice: number;
    tcgdbPokemonPriceChangeMonthlyLowPriceDifference: number;
    tcgdbPokemonPriceChangeMonthlyLowPricePercentage: number;
    tcgdbPokemonPriceChangeMonthlyCurrentMidPrice: number;
    tcgdbPokemonPriceChangeMonthlyPreviousMidPrice: number;
    tcgdbPokemonPriceChangeMonthlyMidPriceDifference: number;
    tcgdbPokemonPriceChangeMonthlyMidPricePercentage: number;
    tcgdbPokemonPriceChangeMonthlyCurrentHighPrice: number;
    tcgdbPokemonPriceChangeMonthlyPreviousHighPrice: number;
    tcgdbPokemonPriceChangeMonthlyHighPriceDifference: number;
    tcgdbPokemonPriceChangeMonthlyHighPricePercentage: number;
    tcgdbPokemonPriceChangeMonthlyCurrentMarketPrice: number;
    tcgdbPokemonPriceChangeMonthlyPreviousMarketPrice: number;
    tcgdbPokemonPriceChangeMonthlyMarketPriceDifference: number;
    tcgdbPokemonPriceChangeMonthlyMarketPricePercentage: number;
    tcgdbPokemonPriceChangeMonthlySubTypeName: string;
}